llm_service/requirements.txt
pyyaml
fastapi
pydantic
gpt4all
# https://fastapi.tiangolo.com/deployment/manually/
uvicorn[standard]
